+++ Units/UnitFactory.hpp
@@ -15,12 +15,12 @@
 
 class UnitFactory : public std::enable_shared_from_this<UnitFactory> {
 public:
-    virtual std::shared_ptr<TownHall> getTownHall(int x, int y) const = 0;
-    virtual std::shared_ptr<Barracks> getBarracks(int x, int y) const = 0;
-    virtual std::shared_ptr<Worker> getWorker(int x, int y) const = 0;
-    virtual std::shared_ptr<Warrior> getWarrior(int x, int y) const = 0;
-    virtual std::shared_ptr<Archer> getArcher(int x, int y) const = 0;
-    virtual std::shared_ptr<Healer> getHealer(int x, int y) const = 0;
+    virtual std::shared_ptr<TownHall> getTownHall(size_t x, size_t y) const = 0;
+    virtual std::shared_ptr<Barracks> getBarracks(size_t x, size_t y) const = 0;
+    virtual std::shared_ptr<Worker> getWorker(size_t x, size_t y) const = 0;
+    virtual std::shared_ptr<Warrior> getWarrior(size_t x, size_t y) const = 0;
+    virtual std::shared_ptr<Archer> getArcher(size_t x, size_t y) const = 0;
+    virtual std::shared_ptr<Healer> getHealer(size_t x, size_t y) const = 0;
 };
